create ajax add task and set task and foldes

+++ tpl/index_tpl.php
@@ -22,9 +22,7 @@
 </head>
 
 <body>
-    <?php if (!empty($_SESSION['error'])) : ?>
-			<h3 class="session-alert"><?= $_SESSION['error'] ?></h3>
-		<?php unset($_SESSION['error']);endif; ?>
+
   <!-- Start your project here-->
   <section class="vh-100">
     <div class="container py-5 h-100">
@@ -53,21 +51,20 @@
               <!-- start folders -->
               <section class="container" >
                   <div class="row ">
-                    <div class="col-xs-2 col-s-2 col-md-2 col-l-2 col-xl-2 active" >
+                    <div class="col-xs-2 col-s-2 col-md-2 col-l-2 col-xl-2 <?= empty($_GET['folder_id']) ? 'active' : '' ?>" >
                       <a href="?folder_id=0">
                         <i class="fa fa-folder folders"></i>
-                        <p class="name-folder ">Home folder</p>
-                        
+                        <p class="name-folder ">All folder</p>
                       </a>
                     </div>
                   </div>
                 <?php foreach($folder_data as $key => $valuse): ?>
                   <div class="row ">
-                    <div class="col-xs-2 col-s-2 col-md-2 col-l-2 col-xl-2" >
+                    <div class="col-xs-2 col-s-2 col-md-2 col-l-2 col-xl-2  <?= !empty($_GET['folder_id']) && $_GET['folder_id'] ==$valuse['id'] ? 'active' : ''; ?>" >
                       <a href="?folder_id=<?= $valuse['id']; ?>">
                         <i class="fa fa-folder folders"></i>
                         <p class="name-folder"><?= $valuse['name']; ?></p>
-                        <a href="?delete_folder=<?= $valuse['id']; ?>"><i class="fas fa-trash-alt delete_folder"></i></a>
+                        <a href="?delete_folder=<?= $valuse['id'] ; ?>" onclick="return confirm('are you sure to delete folder name = <?= $valuse['name']; ?> ?');"><i class="fas fa-trash-alt delete_folder"></i></a>
                       </a>
                     </div>
                   </div>
@@ -93,28 +90,32 @@
               </div>   
               <!-- end add task -->
               <div class="task-list">
-              <?php foreach ($task_data as $key => $value) : ?>
-                <ul class="list-group list-group-horizontal rounded-0 bg-transparent">
-                  <li class="list-group-item d-flex align-items-center ps-0 pe-3 py-1 rounded-0 border-0 bg-transparent">
-                    <div class="form-check">
-                      <input class="form-check-input me-0" type="checkbox" value="" id="flexCheckChecked1" aria-label="..." class ="<?php $value['is_done'] ? 'checked'  : '' ; ?>" />
-                    </div>
-                  </li>
-                  <li class="list-group-item px-3 py-1 d-flex align-items-center flex-grow-1 border-0 bg-transparent">
-                    <p class="lead fw-normal mb-0" id="title_list"><?= $value['title'] ;?></p>
-                  </li>
-                  <li class="list-group-item ps-3 pe-0 py-1 rounded-0 border-0 bg-transparent">
-                    <div class="d-flex flex-row justify-content-end mb-1">
-                      <a href="#!" class="text-info" data-mdb-toggle="tooltip" title="Edit todo"><i class="fas fa-pencil-alt me-3"></i></a>
-                      <a href="?delete_task=<?= $value['id'] ?>" class="text-danger" data-mdb-toggle="tooltip" title="Delete todo"><i class="fas fa-trash-alt"></i></a>
-                    </div>
-                    <div class="text-end text-muted">
-                      <a href="#!" class="text-muted" data-mdb-toggle="tooltip" title="Created date">
-                        <p class="small mb-0"><i class="fas fa-info-circle me-2"></i><?= $value['created_at'] ; ?></p></a>
-                    </div>
-                  </li>
-                </ul>
-              <?php endforeach ; ?>
+              <?php if( sizeof($task_data) > 0 ): ?>
+                <?php foreach ($task_data as $key => $value) : ?>
+                  <ul class="list-group list-group-horizontal rounded-0 bg-transparent">
+                    <li class="list-group-item d-flex align-items-center ps-0 pe-3 py-1 rounded-0 border-0 bg-transparent">
+                      <div class="form-check">
+                        <input class="form-check-input me-0" type="checkbox" value="" id="flexCheckChecked1" aria-label="..." class ="<?php $value['is_done'] ? 'checked'  : '' ; ?>" />
+                      </div>
+                    </li>
+                    <li class="list-group-item px-3 py-1 d-flex align-items-center flex-grow-1 border-0 bg-transparent">
+                      <p class="lead fw-normal mb-0" id="title_list"><?= $value['title'] ;?></p>
+                    </li>
+                    <li class="list-group-item ps-3 pe-0 py-1 rounded-0 border-0 bg-transparent">
+                      <div class="d-flex flex-row justify-content-end mb-1">
+                        <a href="#!" class="text-info" data-mdb-toggle="tooltip" title="Edit todo"><i class="fas fa-pencil-alt me-3"></i></a>
+                        <a href="?delete_task=<?= $value['id'] ; ?>" class="text-danger" data-mdb-toggle="tooltip" title="Delete todo" onclick="return confirm('are you sure to delete task ?');"><i class="fas fa-trash-alt"></i></a>
+                      </div>
+                      <div class="text-end text-muted">
+                        <a href="#!" class="text-muted" data-mdb-toggle="tooltip" title="Created date">
+                          <p class="small mb-0"><i class="fas fa-info-circle me-2"></i><?= $value['created_at'] ; ?></p></a>
+                      </div>
+                    </li>
+                  </ul>
+                <?php endforeach ; ?>
+              <?php else : ?>
+                <p>folder is empty please add task ... </p>
+              <?php endif ; ?>
               </div>
 
             </div>
@@ -134,7 +135,6 @@
 
   <script>
     $(document).ready(function(){
-      
       //add folder to page  
       $('#addBtnNewFolder').click(function(e){
         var input = $('input#addNewFolderInput') ;
@@ -144,7 +144,7 @@
           data : {action:'addFolder',folderName:input.val()},
           success : function(response){
             if (response=='1') {
-              $('<div class="row "> <div class="col-xs-2 col-s-2 col-md-2 col-l-2 col-xl-2" > <a href="?folder_id=<?= $valuse['id']; ?>"> <i class="fa fa-folder folders"></i> <p class="name-folder">' +input.val()+ '</p> <a href="?delete_folder=<?= $valuse['id']; ?>"></a> </a> </div> </div>').appendTo('.container')
+              location.reload() ;
             }else{
               alert(response) ;
             }
@@ -152,21 +152,21 @@
         });
       }) ;
        //add task to page  
-      //  $('#addBtnNewTask').click(function(e){
-      //   var inputTask = $('input#addTaskNewInput') ;
-      //   $.ajax({
-      //     url : 'process/ajax_handler.php',
-      //     method : 'post',
-      //     data : {action:'addTask',taskName:inputTask.val()},
-      //     success : function(response){
-      //       if (response=='1') {
-      //         $('<ul class="list-group list-group-horizontal rounded-0 bg-transparent"> <li class="list-group-item d-flex align-items-center ps-0 pe-3 py-1 rounded-0 border-0 bg-transparent"> <div class="form-check"> <input class="form-check-input me-0" type="checkbox" value="" id="flexCheckChecked1" aria-label="..." " /> </div> </li> <li class="list-group-item px-3 py-1 d-flex align-items-center flex-grow-1 border-0 bg-transparent"> <p class="lead fw-normal mb-0" id="title_list">'  +input.val()+ '</p> </li> <li class="list-group-item ps-3 pe-0 py-1 rounded-0 border-0 bg-transparent"> <div class="d-flex flex-row justify-content-end mb-1"> <a href="#!" class="text-info" data-mdb-toggle="tooltip" title="Edit todo"><i class="fas fa-pencil-alt me-3"></i></a> <a href="?delete_task=<?= $value['id'] ?>" class="text-danger" data-mdb-toggle="tooltip" title="Delete todo"><i class="fas fa-trash-alt"></i></a> </div> <div class="text-end text-muted"> <a href="#!" class="text-muted" data-mdb-toggle="tooltip" title="Created date"> <p class="small mb-0"><i class="fas fa-info-circle me-2"></i><?= $value['created_at'] ; ?></p></a> </div> </li> </ul>').appendTo('.task-list')
-      //       }else{
-      //         alert(response) ;
-      //       }
-      //     }
-      //   });
-      // }) ;
+       $('#addBtnNewTask').click(function(e){
+        var inputTask = $('input#addTaskNewInput') ;
+        $.ajax({
+          url : 'process/ajax_handler.php',
+          method : 'post',
+          data : {action:'addTask',folderId : <?= $_GET['folder_id']; ?>,taskTitle:inputTask.val()},
+          success : function(response){
+            if (response=='1') {
+              location.reload() ;
+            }else{
+              alert(response) ;
+            }
+          }
+        });
+      }) ;
     }) ;
   </script>
 </body>
